Modularize BrightnessControl
