+++ src/daemon/core/core.rs
@@ -32,530 +32,29 @@ use regex::Regex;
 
 use futures::stream::FuturesOrdered;
 
-pub const SOCKET_PATH: &str = "/tmp/brightness_control_socket.sock";
+use crate::daemon::fs::*;
 
-pub const CONFIG_TEMPLATE: &str = include_str!("../config_template.toml");
-
-lazy_static! {
-    static ref DEFAULT_CONFIG: DaemonOptions = {
-        let parsed_toml: DaemonOptions = toml::from_slice(CONFIG_TEMPLATE.as_bytes()).unwrap();
-        parsed_toml
-    };
-
-    pub static ref BINCODE_OPTIONS: DefaultOptions = {
-        let options = bincode::DefaultOptions::default();
-        options.with_fixint_encoding();
-        options
-    };
-
-    static ref XRANDR_DISPLAY_INFORMATION_REGEX: Regex = {
-        Regex::new(r"(?x) # ignore whitespace
-        # [[:alpha:]] represents ascii letters
-        ^([[:alpha:]]+-[[:digit:]]+) # 0 : the adapter name
-        \ # space
-        connected
-        \ # space
-        .*? # optional other words
-        ([[:digit:]]+) # 1 : width
-        x
-        ([[:digit:]]+) # 2 : height
-        \+
-        ([[:digit:]]+) # 3 : x_offset
-        \+
-        ([[:digit:]]+) # 4 : y_offset
-        ").unwrap()
-    };
-}
-
-// encapsulates information from xrandr --current
-struct Monitor {
-    adapter_name: String,
-    width: u32,
-    height: u32,
-    x_offset: u32,
-    y_offset: u32
-}
-
-impl Monitor {
-    fn new(xrandr_line: &str) -> Option<Monitor> {
-        // eDP-1 connected primary 1920x1080+0+0 (normal left inverted right x axis y axis) 344mm x 193mm
-        // HDMI-1 connected 1280x1024+1920+28 (normal left inverted right x axis y axis) 338mm x 270mm
-        // <adapter> connected [primary] <width>x<height>+<x offset>+<y offset> (<flags>) <something>mm x <something else>mm
-        let captures = (*XRANDR_DISPLAY_INFORMATION_REGEX).captures(xrandr_line);
-
-        if let Some(captures) = captures {
-            // 0 points to the entire match, so skip
-            let adapter_name = captures.get(1).unwrap().as_str().to_owned();
-
-            let mut parse_int = | num: regex::Match | {
-                num.as_str().parse::<u32>()
-            };
-
-            (|| {
-                let width = parse_int(captures.get(2).unwrap())?;
-                let height = parse_int(captures.get(3).unwrap())?;
-                let x_offset = parse_int(captures.get(4).unwrap())?;
-                let y_offset = parse_int(captures.get(5).unwrap())?;
-
-                std::result::Result::<Option<Monitor>, std::num::ParseIntError>::Ok(
-                    Some(
-                        Monitor {
-                            adapter_name,
-                            width,
-                            height,
-                            x_offset,
-                            y_offset
-                        }
-                    )
-                )
-            })().unwrap_or(None)
+use super::{
+    display::*,
+    super::{
+        fs::*,
+        config::{
+            runtime::*,
+            persistent::*
+        },
+        util::{
+            lock::*,
+            io::*
+        },
+        super::{
+            shared::*
         }
-        else {
-            None
-        }
-    }
-}
-
-struct SocketMessageHolder {
-    // queue messages into this and then push them all to the socket at the end
-    messages: Vec<SocketMessage>,
-    socket: UnixStream
-}
-
-impl SocketMessageHolder {
-    fn new(socket: UnixStream) -> SocketMessageHolder {
-        SocketMessageHolder {
-            messages: Vec::with_capacity(5),
-            socket
-        }
-    }
-
-    fn queue_message<T>(&mut self, message: T, log_socket_error: bool)
-    where T: Into<String> {
-        self.messages.push(
-            SocketMessage {
-                message: message.into(),
-                log_socket_error
-            }
-        )
-    }
-
-    fn queue_success<T>(&mut self, message: T)
-    where T: Into<String> {
-        self.queue_message(message, true)
-    }
-
-    fn queue_error<T>(&mut self, message: T)
-    where T: Into<String> {
-        self.queue_message(message, false)
-    }
-
-    // NOTE remember to consume before it goes out of scope
-    fn consume(mut self) {
-        // write all messages to the socket
-        tokio::spawn(
-            async move {
-                for message_struct in self.messages.into_iter() {
-                    let message = message_struct.message;
-                    if let Err(e) = self.socket.write_all(&message.as_bytes()).await {
-                        if message_struct.log_socket_error {
-                            eprintln!("Failed to write \"{}\" to socket: {}", message, e);
-                        }
-                    }
-                }
-
-                // cleanup; close connection
-                let _ = self.socket.shutdown(std::net::Shutdown::Both);
-            }
-        );
-    }
-}
-
-struct SocketMessage {
-    message: String,
-    log_socket_error: bool
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct BrightnessInput {
-    pub brightness: Option<BrightnessChange>,
-    pub override_fade: Option<bool>,
-    pub terminate_fade: bool
-}
-
-impl BrightnessInput {
-    fn is_active(&self) -> bool {
-        self.brightness.is_some() || self.terminate_fade
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub enum BrightnessChange {
-    Adjustment(i8),
-    Set(u8)
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub enum GetProperty {
-    Brightness,
-    IsFading,
-    Mode,
-    Displays,
-    Config
-}
+    },
+};
 
 enum ProcessInputExitCode {
     Normal,
     Shutdown
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ProgramInput {
-    brightness: BrightnessInput,
-    get_property: Option<GetProperty>,
-    configure_display: bool,
-    toggle_nightlight: bool,
-    reload_configuration: bool,
-    shutdown: bool
-}
-
-impl ProgramInput {
-    pub fn new(brightness: BrightnessInput, get_property: Option<GetProperty>, configure_display: bool, toggle_nightlight: bool, reload_configuration: bool, shutdown: bool) -> ProgramInput {
-        ProgramInput {
-            brightness,
-            get_property,
-            configure_display,
-            toggle_nightlight,
-            reload_configuration,
-            shutdown,
-        }
-    }
-
-    pub fn returns_feedback(&self) -> bool {
-        // create a vector of all options that send back feedback to the client
-        let feedback_returning_options = vec![
-            self.reload_configuration
-        ];
-
-        // returns true if any of them are true
-        feedback_returning_options.iter().any(|&b| b)
-    }
-}
-
-struct FileUtils {
-    project_directory: ProjectDirs,
-    file_open_options: OpenOptions,
-}
-
-impl FileUtils {
-    fn new() -> Result<FileUtils> {
-        let project_directory = get_project_directory()?;
-
-        let file_open_options = {
-            let mut file_open_options = OpenOptions::new();
-            file_open_options.read(true);
-            file_open_options.write(true);
-            file_open_options.create(true);
-            file_open_options
-        };
-
-        Ok(
-            FileUtils {
-                project_directory,
-                file_open_options
-            }
-        )
-    }
-
-    // returns the file and whether or not it existed prior to opening it
-    async fn open_configuration_file(&self) -> Result<(File, bool)> {
-        let config_dir = self.project_directory.config_dir();
-        let filepath = config_dir.join("config.toml");
-
-        if !config_dir.exists() {
-            fs::create_dir(self.project_directory.config_dir()).await?;
-        }
-
-        let file_exists = fs::metadata(&filepath).await.is_ok();
-        Ok( (self.file_open_options.open(filepath).await?, file_exists) )
-    }
-
-    async fn open_cache_file_with_options(&self, file_name: &str, open_options: &OpenOptions) -> Result<File> {
-        let filepath = self.project_directory.cache_dir().join(file_name);
-        open_options.open(filepath).await
-    }
-
-    async fn open_cache_file(&self, file_name: &str) -> Result<File> {
-        self.open_cache_file_with_options(file_name, &self.file_open_options).await
-    }
-
-    async fn get_mode_file(&self) -> Result<File> {
-        self.open_cache_file("mode").await
-    }
-
-    async fn get_brightness_file(&self) -> Result<File> {
-        self.open_cache_file("brightness").await
-    }
-
-    // 0 for regular
-    // 1 for night light
-    // gets the mode written to disk; if invalid, writes a default and returns it
-    async fn get_written_mode(&self) -> Result<bool> {
-        let mut mode_file = self.get_mode_file().await?;
-        mode_file.set_len(1).await?;
-
-        get_valid_data_or_write_default(&mut mode_file, &| data_in_file: &String | {
-            if let Ok(num) = data_in_file.parse::<u8>() {
-                if num == 0 || num == 1 {
-                    // cannot cast "num as bool" normally
-                    return Ok(num != 0);
-                }
-            }
-
-            return Err(Error::new(ErrorKind::InvalidData, "Invalid mode"));
-
-        }, false).await
-    }
-
-    async fn get_written_brightness(&self) -> Result<f64> {
-        let mut brightness_file = self.get_brightness_file().await?;
-
-        get_valid_data_or_write_default(&mut brightness_file, &| data_in_file: &String | {
-            // need to trim this because the newline character breaks the parse
-            if let Ok(num) = data_in_file.trim_end().parse::<f64>() {
-                // check bounds; don't allow 0.0 brightness
-                if num <= 100.0 && num > 0.0 {
-                    return Ok(num);
-                }
-            }
-
-            return Err(Error::new(ErrorKind::InvalidData, "Invalid brightness"));
-        }, 100.0).await
-    }
-
-    async fn write_brightness(&self, brightness: f64) -> Result<()> {
-        let mut brightness_file = self.get_brightness_file().await?;
-        overwrite_file_with_content(&mut brightness_file, brightness).await?;
-        Ok(())
-    }
-
-    async fn write_mode(&self, mode: bool) -> Result<()> {
-        let mut mode_file = self.get_mode_file().await?;
-        overwrite_file_with_content(&mut mode_file, mode).await?;
-        Ok(())
-    }
-
-    // checks if the header in the config template reads a different version
-    // than the current version of BrightnessControl
-    // if they do not match OR anything goes wrong during the check, overwrites
-    // the template with CONFIG_TEMPLATE
-    async fn update_config_template(&self) -> Result<()> {
-        // read header of current template
-        let data_dir = self.project_directory.data_dir();
-
-        if !data_dir.exists() {
-            fs::create_dir_all(data_dir).await?;
-        }
-
-        let template_filepath = data_dir.join("config_template.toml");
-        let template_exists = template_filepath.exists();
-
-        let mut template_file = self.file_open_options.open(&template_filepath).await?;
-        let mut template_file_clone = template_file.try_clone().await?;
-
-        // this allows us to handle any errors in this section the same way
-        let overwrite_template_result = (|| async move {
-            if template_exists {
-                // we only have to read until the first newline
-                // # a.b.c
-                // 12345678
-                // 2 bytes extra incase 'a' becomes double(/triple)-digits
-                const BYTES_TO_READ: usize = 20;
-
-                let buffered_reader = BufReader::with_capacity(BYTES_TO_READ, &mut template_file);
-
-                // buffered_reader.fill_buf().await?;
-
-                if let Ok(Some(first_line)) = buffered_reader.lines().next_line().await {
-                    const NUM_CHARS_TO_IGNORE: usize = "# v".len();
-                    // Strings from BufReader::lines do not include newlines at the
-                    // end
-                    let version_string = &first_line[NUM_CHARS_TO_IGNORE..];
-
-                    let current_version_string = {
-                        let beginning_trimmed = &CONFIG_TEMPLATE[NUM_CHARS_TO_IGNORE..];
-                        let newline_index = beginning_trimmed.find("\n").unwrap();
-                        &beginning_trimmed[..newline_index]
-                    };
-
-                    // compare to actual version string
-                    if version_string.eq(current_version_string) {
-                        return Ok(());
-                    }
-                    else {
-                        println!("Config template updated! \"{}\" to \"{}\"", version_string, current_version_string);
-                    }
-                }
-            }
-            else {
-                println!("Config template saved to {}", &template_filepath.display());
-            }
-
-            // the cause of this error is irrelevant, so it doesnt need a
-            // message
-            Err(Error::new(ErrorKind::Other, ""))
-        })().await;
-
-        // dont care about the cause
-        if let Err(_) = overwrite_template_result {
-            // overwrite
-            overwrite_file_with_content(&mut template_file_clone, CONFIG_TEMPLATE).await?;
-        }
-
-        Ok(())
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-struct NightlightOptions {
-    xrandr_gamma: String,
-    redshift_temperature: u32
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-struct FadeOptions {
-    threshold: u8,
-    // milliseconds
-    total_duration: u32,
-    // milliseconds
-    step_duration: u32
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-struct DaemonOptions {
-    use_redshift: bool,
-    auto_remove_displays: bool,
-    fade_options: FadeOptions,
-    nightlight_options: NightlightOptions
-}
-
-impl DaemonOptions {
-    fn default() -> DaemonOptions {
-        (*DEFAULT_CONFIG).clone()
-    }
-}
-
-struct BrightnessState {
-    // receiver end of channel in mutex
-    brightness: NonReadBlockingRWLock<f64, mpsc::UnboundedReceiver<(BrightnessInput, SocketMessageHolder)>>,
-    fade_notifier: mpsc::UnboundedSender<(BrightnessInput, SocketMessageHolder)>,
-    is_fading: Cell<bool>
-}
-
-unsafe impl Send for BrightnessState {}
-unsafe impl Sync for BrightnessState {}
-
-impl BrightnessState {
-    fn new(initial_brightness: f64) -> BrightnessState {
-        let (tx, rx) = mpsc::unbounded_channel::<(BrightnessInput, SocketMessageHolder)>();
-
-        BrightnessState {
-            brightness: NonReadBlockingRWLock::new(initial_brightness, rx),
-            fade_notifier: tx,
-            is_fading: Cell::new(false)
-        }
-    }
-
-    fn get(&self) -> f64 {
-        self.brightness.get()
-    }
-
-    fn get_fade_notifier(&self) -> mpsc::UnboundedSender<(BrightnessInput, SocketMessageHolder)> {
-        self.fade_notifier.clone()
-    }
-
-    fn try_lock_brightness<'a>(&'a self) -> Option<MutexGuardRefWrapper<'a, f64, mpsc::UnboundedReceiver<(BrightnessInput, SocketMessageHolder)>>> {
-        self.brightness.try_lock_mut()
-    }
-
-    async fn lock_brightness<'a>(&'a self) -> MutexGuardRefWrapper<'a, f64, mpsc::UnboundedReceiver<(BrightnessInput, SocketMessageHolder)>> {
-        self.brightness.lock_mut().await
-    }
-}
-
-struct MutexGuardRefWrapper<'a, T: Copy, K> {
-    internal: &'a mut T,
-    mutex_guard: MutexGuard<'a, K>
-}
-
-impl<'a, T, K> MutexGuardRefWrapper<'a, T, K>
-where T: Copy {
-    fn set(&mut self, new_value: T) {
-        *self.internal = new_value;
-    }
-}
-
-// a version of RWLock that does not block readers from reading while a writer writes
-// by providing them with a copy of the internal value
-// this is meant to be used with primitives which have cheap Copy implementations
-struct NonReadBlockingRWLock<T: Copy, K> {
-    internal: Cell<T>,
-    write_mutex: Mutex<K>
-}
-
-unsafe impl<T, K> Send for NonReadBlockingRWLock<T, K>
-where T: Copy {}
-
-unsafe impl<T, K> Sync for NonReadBlockingRWLock<T, K>
-where T: Copy {}
-
-impl<T, K> NonReadBlockingRWLock<T, K>
-where T: Copy {
-    fn new(initial_value: T, mutex_value: K) -> NonReadBlockingRWLock<T, K> {
-        NonReadBlockingRWLock {
-            internal: Cell::new(initial_value),
-            write_mutex: Mutex::new(mutex_value)
-        }
-    }
-
-    fn get(&self) -> T {
-        self.internal.get()
-    }
-
-    async fn set_value(&self, new_value: T) {
-        self.write_mutex.lock().await;
-        self.internal.set(new_value);
-    }
-
-    fn try_lock_mut<'a>(&'a self) -> Option<MutexGuardRefWrapper<'a, T, K>> {
-        let guard = self.write_mutex.try_lock();
-
-        if let Ok(guard) = guard {
-            Some(
-                MutexGuardRefWrapper {
-                    // need to get a mutable reference to internal without making the
-                    // function take a mutable reference
-                    internal: unsafe { self.internal.as_ptr().as_mut().unwrap() },
-                    mutex_guard: guard
-                }
-            )
-        }
-        else {
-            None
-        }
-    }
-    
-
-    async fn lock_mut<'a>(&'a self) -> MutexGuardRefWrapper<'a, T, K> {
-        let guard = self.write_mutex.lock().await;
-
-        MutexGuardRefWrapper {
-            // need to get a mutable reference to internal without making the
-            // function take a mutable reference
-            internal: unsafe { self.internal.as_ptr().as_mut().unwrap() },
-            mutex_guard: guard
-        }
-    }
 }
 
 // separate RwLocks so they can be modified concurrently
@@ -938,7 +437,7 @@ impl Daemon {
                     let config_result = get_configuration_from_file(&mut configuration_file);
 
                     match config_result.await {
-                        Ok(config) => {
+                        std::result::Result::<DaemonOptions, toml::de::Error>::Ok(config) => {
                             *self.config.write().await = config;
 
                             socket_holder.queue_success("Successfully reloaded configuration!");
@@ -1209,130 +708,6 @@ impl Daemon {
     }
 }
 
-async fn overwrite_file_with_content<T>(file: &mut File, new_content: T) -> Result<()>
-where T: Display {
-    file.seek(std::io::SeekFrom::Start(0)).await?;
-
-    let formatted_new_content = format!("{}", new_content);
-
-    // <<NOTE>> this can overflow? len() returns a usize
-    file.set_len(formatted_new_content.len() as u64).await?;
-
-    file.write_all(&formatted_new_content.as_bytes()).await?;
-
-    Ok(())
-}
-
-// where ....
-async fn get_valid_data_or_write_default<T>(file: &mut File, data_validator: &dyn Fn(&String) -> Result<T>, default_value: T) -> Result<T>
-where T: Display {
-    let mut file_clone = file.try_clone().await?;
-
-    let file_contents = {
-        // wrapping in a closure allows the inner else and the
-        // outer else clauses to share the same code
-        // here, we want to return None if the file does not exist
-        // or if the file's contents are not readable as a number
-        (|| async move {
-            let mut buffer: Vec<u8> = Vec::new();
-            file.read_to_end(&mut buffer).await?;
-
-            let string = unsafe {
-                String::from_utf8_unchecked(buffer)
-            };
-
-            data_validator(&string)
-        })()
-    }.await;
-
-    if let Ok(contents) = file_contents {
-        Ok(contents)
-    }
-    else {
-        overwrite_file_with_content(&mut file_clone, &default_value).await?;
-        Ok(default_value)
-    }
-}
-
-fn get_project_directory() -> Result<directories::ProjectDirs> {
-    let project_directory = ProjectDirs::from("", "Sridaran Thoniyil", "BrightnessControl");
-    // did not use if let because it would require the entire function to be indented
-
-    if let None = project_directory {
-        panic!("Cannot find base directory");
-    }
-
-    let project_directory = project_directory.unwrap();
-    // cache the mode
-    let cache_directory = project_directory.cache_dir();
-
-    if !cache_directory.exists() {
-        std::fs::create_dir_all(cache_directory)?;
-    }
-
-    Ok(project_directory)
-}
-
-async fn get_current_connected_displays() -> Result<Vec<Monitor>> {
-    let mut xrandr_current = Command::new("xrandr");
-    xrandr_current.arg("--current");
-    let command_output = xrandr_current.output().await?;
-    // the '&' operator dereferences ascii_code so that it can be compared with a regular u8
-    // its original type is &u8
-    let output_lines = command_output.stdout.split(| &ascii_code | ascii_code == '\n' as u8);
-
-    let connected_displays: Vec<Monitor> = output_lines.filter_map(|line| {
-        // if valid UTF-8, pass to Monitor
-        if let Ok(line) = std::str::from_utf8(line) {
-            Monitor::new(line)
-        }
-        else {
-            None
-        }
-    }).collect();
-
-    Ok(connected_displays)
-}
-
-async fn get_configuration_from_file(configuration_file: &mut File) -> std::result::Result<DaemonOptions, toml::de::Error> {
-    // 8 KB
-    const INITIAL_BUFFER_SIZE: usize = 8 * 1024;
-
-    let mut configuration_buffer = Vec::with_capacity(INITIAL_BUFFER_SIZE);
-
-    // fill buffer
-    if let Err(e) = configuration_file.read_to_end(&mut configuration_buffer).await {
-        eprintln!("Failed to read from configuration file! {}", e);
-    }
-
-    let parsed_toml: toml::Value = toml::from_slice(&configuration_buffer[..configuration_buffer.len()])?;
-
-    let mut config = DaemonOptions::default();
-
-    macro_rules! overwrite_values {
-        ( $( $x:ident ),* ) => {
-            {
-                $(
-                    if let Some(option) = parsed_toml.get(stringify!($x)) {
-                        config.$x = option.clone().try_into()?;
-                    }
-                )*
-            }
-        };
-    }
-
-    // TODO figure out how to use derive macro for this
-    overwrite_values!(use_redshift, auto_remove_displays, fade_options, nightlight_options);
-
-    return Ok(config);
-}
-
-async fn configure_displays() -> Result<Vec<Monitor>> {
-    let connected_displays = get_current_connected_displays().await?;
-
-    Ok(connected_displays)
-}
-
 fn register_sigterm_handler() -> Result<()> {
     unsafe {
         signal_hook::register(signal_hook::SIGTERM, move || {
@@ -1386,7 +761,7 @@ fn register_sigterm_handler() -> Result<()> {
 pub fn daemon(fork: bool) -> Result<()> {
     let file_utils = FileUtils::new()?;
 
-    let pid_file_path = &file_utils.project_directory.cache_dir().join("daemon.pid");
+    let pid_file_path = file_utils.get_cache_dir().join("daemon.pid");
 
     let cache_dir = file_utils.project_directory.cache_dir();
