+++ src/daemon/mod.rs
@@ -1,0 +1,8 @@
+pub mod core;
+pub mod fs;
+pub mod config;
+pub mod util;
+
+pub fn start_daemon(fork: bool) -> std::io::Result<()> {
+    core::core::daemon(fork)
+}
