+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod daemon;
 pub mod client;
+pub mod shared;
